Move sanitizeFilename to shared
